+++ src/affine.rs
@@ -513,6 +513,33 @@ where
         }
     }
 
+    /// Creates an affine transform from a column-major array of elements.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use ggmath::{Affine2, Vec2};
+    /// #
+    /// let affine = Affine2::from_column_array(&[1.0, 2.0, 3.0, 4.0, 5.0, 6.0]);
+    /// assert_eq!(
+    ///     affine,
+    ///     Affine2::from_columns(&[
+    ///         Vec2::new(1.0, 2.0),
+    ///         Vec2::new(3.0, 4.0),
+    ///         Vec2::new(5.0, 6.0),
+    ///     ]),
+    /// );
+    /// ```
+    #[inline]
+    #[must_use]
+    pub const fn from_column_array(array: &[T; 6]) -> Self {
+        Self::from_columns(&[
+            Vector::<2, T, A>::new(array[0], array[1]),
+            Vector::<2, T, A>::new(array[2], array[3]),
+            Vector::<2, T, A>::new(array[4], array[5]),
+        ])
+    }
+
     /// Returns a reference to the affine transform's columns.
     #[inline]
     #[must_use]
@@ -660,6 +687,34 @@ where
         }
     }
 
+    /// Creates an affine transform from a column-major array of elements.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use ggmath::{Affine2, Vec2};
+    /// #
+    /// let affine = Affine2::from_column_array(&[1.0, 2.0, 3.0, 4.0, 5.0, 6.0]);
+    /// assert_eq!(
+    ///     affine,
+    ///     Affine2::from_columns(&[
+    ///         Vec2::new(1.0, 2.0),
+    ///         Vec2::new(3.0, 4.0),
+    ///         Vec2::new(5.0, 6.0),
+    ///     ]),
+    /// );
+    /// ```
+    #[inline]
+    #[must_use]
+    pub const fn from_column_array(array: &[T; 12]) -> Self {
+        Self::from_columns(&[
+            Vector::<3, T, A>::new(array[0], array[1], array[2]),
+            Vector::<3, T, A>::new(array[3], array[4], array[5]),
+            Vector::<3, T, A>::new(array[6], array[7], array[8]),
+            Vector::<3, T, A>::new(array[9], array[10], array[11]),
+        ])
+    }
+
     /// Returns a reference to the affine transform's columns.
     #[inline]
     #[must_use]
@@ -767,6 +822,35 @@ where
                 array[0], array[1], array[2], array[3], array[4],
             ))
         }
+    }
+
+    /// Creates an affine transform from a column-major array of elements.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use ggmath::{Affine2, Vec2};
+    /// #
+    /// let affine = Affine2::from_column_array(&[1.0, 2.0, 3.0, 4.0, 5.0, 6.0]);
+    /// assert_eq!(
+    ///     affine,
+    ///     Affine2::from_columns(&[
+    ///         Vec2::new(1.0, 2.0),
+    ///         Vec2::new(3.0, 4.0),
+    ///         Vec2::new(5.0, 6.0),
+    ///     ]),
+    /// );
+    /// ```
+    #[inline]
+    #[must_use]
+    pub const fn from_column_array(array: &[T; 20]) -> Self {
+        Self::from_columns(&[
+            Vector::<4, T, A>::new(array[0], array[1], array[2], array[3]),
+            Vector::<4, T, A>::new(array[4], array[5], array[6], array[7]),
+            Vector::<4, T, A>::new(array[8], array[9], array[10], array[11]),
+            Vector::<4, T, A>::new(array[12], array[13], array[14], array[15]),
+            Vector::<4, T, A>::new(array[16], array[17], array[18], array[19]),
+        ])
     }
 
     /// Returns a reference to the affine transform's columns.
@@ -1633,6 +1717,44 @@ mod tests {
                     ]),
                     Vector::<4, T, A>::new(m, n, o, p)
                 )
+            );
+        });
+    }
+
+    #[test]
+    fn test_from_column_array() {
+        for_parameters!(|T: PrimitiveNumber, A| {
+            let [x, y, z, w, a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p] =
+                std::array::from_fn(T::as_from);
+
+            assert_eq!(
+                Affine::<2, T, A>::from_column_array(&[x, y, z, w, a, b]),
+                Affine::<2, T, A>::from_columns(&[
+                    Vector::<2, T, A>::new(x, y),
+                    Vector::<2, T, A>::new(z, w),
+                    Vector::<2, T, A>::new(a, b)
+                ])
+            );
+            assert_eq!(
+                Affine::<3, T, A>::from_column_array(&[x, y, z, w, a, b, c, d, e, f, g, h]),
+                Affine::<3, T, A>::from_columns(&[
+                    Vector::<3, T, A>::new(x, y, z),
+                    Vector::<3, T, A>::new(w, a, b),
+                    Vector::<3, T, A>::new(c, d, e),
+                    Vector::<3, T, A>::new(f, g, h)
+                ])
+            );
+            assert_eq!(
+                Affine::<4, T, A>::from_column_array(&[
+                    x, y, z, w, a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p
+                ]),
+                Affine::<4, T, A>::from_columns(&[
+                    Vector::<4, T, A>::new(x, y, z, w),
+                    Vector::<4, T, A>::new(a, b, c, d),
+                    Vector::<4, T, A>::new(e, f, g, h),
+                    Vector::<4, T, A>::new(i, j, k, l),
+                    Vector::<4, T, A>::new(m, n, o, p)
+                ])
             );
         });
     }