+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Affine transform functions:
+  - `from_column_array`
+
 ## [0.16.5] - 2026-04-01
 
 ### Added